+++ src/components/core/data-list.stories.tsx
@@ -3,6 +3,7 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 export default {
   component: DataList,
+  tags: ["docsPage"],
 } as Meta<DataListProps>;
 
 export const Basic: StoryObj<DataListProps> = {