+++ src/fhir/system-urls.ts
@@ -22,6 +22,12 @@ export const SYSTEM_MARITAL_STATUS =
 
 export const SYSTEM_PRACTITIONER_ID = "https://zusapi.com/practitioner_id";
 
+export const SYSTEM_PROVENANCE_ACTIVITY_TYPE =
+  "http://terminology.hl7.org/3.1.0/CodeSystem-v3-DataOperation.html";
+
+export const SYSTEM_PROVENANCE_AGENT_TYPE =
+  "http://terminology.hl7.org/CodeSystem/provenance-participant-type";
+
 export const SYSTEM_RELATIONSHIP =
   "http://terminology.hl7.org/CodeSystem/v2-0131";